src/main/java/com/billme/billme/service/currency/CurrencyService.java
package com.billme.billme.service.currency;

import com.billme.billme.jpa.currency.Currency;
import com.billme.billme.jpa.currency.CurrencyRepository;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.Optional;

@Service
public class CurrencyService {

    private final CurrencyRepository currencyRepository;

    public CurrencyService(CurrencyRepository currencyRepository) {
        this.currencyRepository = currencyRepository;
    }

    @Transactional
    public Optional<Currency> getCurrencyByCode(String code) {
        return currencyRepository.findByCode(code);
    }

}
